Fix build by fixing test errors
